Added memory pool with iterator and allocator
From its documentation:

In situations where objects are frequently allocated and deallocated, using the default
"new" and "delete", which are thinly wrapped _malloc_ and _free_ calls, has two main problems:

1. The allocations can be placed anywhere on the heap,
   giving very poor spatial locality,
   possibly causing a large amount of cache misses, and in turn,
   huge performance losses.
2. Memory fragmentation can occur if these allocations are done
   in conjunction with allocation of other, differently-sized
   objects.

A memory pool is a pattern that can be used to avoid this.
A memory pool is given a fixed maximum size on initialization, at which point
it makes a single, up-fornt allocation for the space for that many objects.
From then on, memory for objects can be divvied out as-needed without making
any actual allocations.
These objects will also have good spatial locality,
given that they are in the same pool.
 *
In order to track which slots in the pool are in use and which aren't,
this class stores a singly-linked list of free slots _inside_ the free slots
(see the Slot union).

At this point, you may be wondering: Isn't there
[boost::pool](http://www.boost.org/doc/libs/1_55_0/libs/pool/doc/html/index.html)?
Yes. But,

1. This allows for iteration over the pool, which `boost::pool` does not.
2. The author has a bit of an aversion to boost, as it often increases compile times
   and is another dependency to worry about. Using boost often feels to the author like
   calling in a carrier battle group when a fishing boat is needed.